Statsutil ready for 0.2 release!
* Fixed a bunch of misc warnings. Mainly unsafe type comparison.
* Rewrote a good part of the item data handling code.
* Wrote the parser / writer  for the item_p and item_s_p files.
* Wrote the XML writer/parser for item data.
* Updated the parsing code to use the new read int function.
* Removed all references to "data_array_struct", as its just unneeded
garbage.
* Added a prototype handler for PMD3 strings. Still need to figure out
how to compute string hashes!
* And a lot more little things..

+++ src/dse/dse_interpreter.cpp
@@ -859,13 +859,13 @@ namespace DSE
             else 
             {
                 //Make several pauses in a row!
-                int nbpauses = ( delta / std::numeric_limits<uint16_t>::max() );
+                unsigned long long nbpauses = delta / std::numeric_limits<uint16_t>::max();
                 if( ( delta % std::numeric_limits<uint16_t>::max() ) != 0 )
                     ++nbpauses;
 
                 unsigned long long pauseleft = delta; //The nb of ticks to pause for remaining
 
-                for( int i = 0; i < nbpauses; ++nbpauses )
+                for( unsigned long long i = 0; i < nbpauses; ++nbpauses )
                 {
                     if( pauseleft < numeric_limits<uint8_t>::max() )
                     {
